Rename `main` to `run`
By convention, the entry point for the console script is a function called
`main`. This function encompasses all the code currently in the `if __name__ ==
"__main__":` block. But there is already a function called `main`, for the main
loop. So we rename this function to `run`, choosing a short name with similar
meaning.