Fix broken color test by testing the values individually
Testing if the sum of the color values is within range is flawed because
one out of range value might still not exceed the minimum or maximum
value. Multiple out of range values can even pass the test by cancelling
each other out.

This patch tests each value individually and raises an exception on the
first out of range value.